// src/app/_interfaces/loggable.interface.ts
// Generated by https://quicktype.io

export interface LoggableCurrency {
    data: LoggableCurrencyData[];
}

export interface LoggableCurrencyData {
    id:        number;
    Hall:      number;
    Amount:    number;
    Name:      string;
    Shortened: string;
    Type:      string;
    Hash:      string;
}